Rename start module to startedit
