+++ app/src/main/res/navigation/tabs_nav_graph.xml
@@ -12,7 +12,7 @@
 
     <dialog
         android:id="@+id/start_time_entry_dialog"
-        android:name="com.toggl.timer.start.ui.StartTimeEntryDialogFragment"
+        android:name="com.toggl.timer.startedit.ui.StartTimeEntryDialogFragment"
         android:label="StartTimeEntryDialogFragment">
         <deepLink app:uri="@string/deep_link_start_time_entry_dialog" />
     </dialog>
