+++ timer/src/main/java/com/toggl/timer/startedit/domain/StartTimeEntryReducer.kt
@@ -1,4 +1,4 @@
-package com.toggl.timer.start.domain
+package com.toggl.timer.startedit.domain
 
 import com.toggl.architecture.DispatcherProvider
 import com.toggl.architecture.core.Effect
