+++ timer/src/test/java/com/toggl/timer/startedit/domain/StopTimeEntryButtonTappedActionTests.kt
@@ -1,4 +1,4 @@
-package com.toggl.timer.start.domain
+package com.toggl.timer.startedit.domain
 
 import com.toggl.models.domain.Workspace
 import com.toggl.repository.interfaces.TimeEntryRepository
